+++ src/main/java/blackjack/domain/betting/BettingToken.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.betting;
 
 import java.util.Objects;
