+++ src/main/java/blackjack/domain/card/CardDenomination.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum CardDenomination {
 	ACE(1, "A"),
