+++ src/main/java/blackjack/domain/card/CardSuit.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum CardSuit {
 	HEART("하트"),
